add Winding heating
