+++ README.md
@@ -1,13 +1,14 @@
 # xlsx_jobs introduction
-## Module:
+## Extra Module:
 1. npTDMS
 2. openpyxl
+3. pillow
 
 ## Function brief Summary
 1. Write Specif BEMF worksheet
 2. Write Specif Conti Torque Worksheet
 3. Write Specif High Speed Worksheet
-4. Now adding more function ...
+4. add Image to Winding Heating Worksheet (developing)
 ## Contact
 **Name:Sylvain Yu**
